src/main/java/com/wcci/reviews/entities/Review.java
package com.wcci.reviews.entities;

public class Review {
    final private Category category;
    final private String title;
    final private String author;
    final private String text;

    public Review(final Category category, final String title, final String author, final String text) {
        this.category = category;
        this.title = title;
        this.author = author;
        this.text = text;
    }

    public Category getCategory() {
        return category;
    }

    public String getTitle() {
        return title;
    }

    public String getAuthor() {
        return author;
    }

    public String getText() {
        return text;
    }

}
